fix(capture): share canonical forge authority in shell tests
