+++ crates/ctx-history-capture/src/repository_attribution/association.rs
@@ -6,11 +6,38 @@ use ctx_history_core::{
 use serde::de::{Deserialize, Deserializer, MapAccess, SeqAccess, Visitor};
 use serde_json::Value;
 use std::{collections::HashSet, fmt};
-use url::Url;
+use url::{Host, Url};
 
-use super::{identity::canonical_url_authority, shell::bounded_pull_request_association_query};
+use super::shell::bounded_pull_request_association_query;
 
 const MAX_ASSOCIATION_OUTPUT_BYTES: usize = 1024 * 1024;
+
+/// Credential-free canonical authority for a forge URL.
+///
+/// Default transport ports are aliases of the ordinary host spelling while a
+/// nondefault port is part of repository authority. IPv6 literals retain
+/// brackets so a following port remains unambiguous.
+pub(super) fn canonical_url_authority(url: &Url) -> Option<String> {
+    let host = match url.host()? {
+        Host::Domain(host) => host.to_ascii_lowercase(),
+        Host::Ipv4(host) => host.to_string(),
+        Host::Ipv6(host) => format!("[{host}]"),
+    };
+    let port = url
+        .port()
+        .filter(|port| Some(*port) != default_port(url.scheme()));
+    Some(port.map_or(host.clone(), |port| format!("{host}:{port}")))
+}
+
+fn default_port(scheme: &str) -> Option<u16> {
+    match scheme {
+        "http" => Some(80),
+        "https" => Some(443),
+        "ssh" => Some(22),
+        "git" => Some(9418),
+        _ => None,
+    }
+}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) struct UnscopedPullRequestAssociationObservation {
